Create a board from a FEN string.
It uses just the piece positions and ignores all the other data for the
moment. I'm not keen on the from_fen method, it seems too long. It also
does no error handling, I'll have to build that in. Maybe that means the
FEN code belongs in its own class.

Also, starting to see some duplication in the tests - there's lots of
string manipulation I think it would be good to get rid of.