Extract squares method from Board.from_fen to FenInput.squares
Moving the method wholesale from one Board to FenInput. I copied the test over
as well, I imagine one or other test will be changed or deleted soon. I'm not
sure whether the FenInput class should return a string or a Board, so I'll
leave it as is for the moment. I don't like how everything on FenInput is a
class method, so I'll fix that first.