Make #pieces_from_fen an instance method.
This means I don't have to pass the state in as an argument. Not sure about
the name. It's not really getting pieces, but the positions of the pieces.

I think I may have gone far enough with refactoring before considering what
happens if the FEN string is wrong. What happens if they give me a character
I'm not expecting? Or if one of the rows has more or less than 8 entries?