Move towards making Board accept an array.
Looking at the tests for checking the bounds of the board, it became apparent
that string manipulation was going to be a pain. So it seems like a good time
to introduce the squares as an array.

I want to do this a bit at a time to avoid breaking the tests, so I've created
the array_to_string method which turns an array into the appropriate string.
It's purely temporary in its current form, but may eventually form part of the
to_s method.

I've started by making the default_starting_position an array and then
immediately converting it back. I can do this for every place I
create a string instead of an array. Then I can switch the entire system to
use arrays in one commit.