Make arguments to Chess::Board mandatory.
I didn't like the fact I had to optional arguments to Board. I don't see the
need for them to be optional just yet. I've had to create the Board.default
method for a new board in the default starting position and I think this is
clearer than defaulting to that if no args are passed to Board.

I could have made the board take a hash of options and have sensible
defaults. That's arguably the most "Ruby" thing to do, but I'm starting to get
a feeling that doing that leads to large and messy method interfaces. These
values aren't really "options", they're integral to the data of a board. At
the moment there's only 2 arguments, but I imagine there'll be more coming
along. At the very least, I could envision the interface being the same as the
values in a FEN string. At that point, if it becomes unwieldy I'll consider
how to make the Board interface nicer.

There's a gap in my tests here - in order to get the tests passing I've had to
default the from_fen method to alway return white's move. I'll fix that next.