Add promotion piece to #move
This allows pieces to be promoted during a move.

We're now at a point where we could emulate a real-life game. The board can be
set up in any position, pieces can be moved and promotion can occur. All the
rules would have to be enforced by the players.

Next up is fleshing out the rules themselves to make sure only valid games can
be played.